fix(sass): fix sass imports for scoped packages
Also requires `@stencil/sass` fix in v1.3.2
Closes #2521
