+++ src/compiler/sys/resolve/resolve-utils.ts
@@ -49,16 +49,45 @@ export const isStencilCoreImport = (p: string) => p.startsWith('@stencil/core');
 
 export const shouldFetchModule = (p: string) => IS_FETCH_ENV && !IS_NODE_ENV && isNodeModulePath(p);
 
-export const isNodeModulePath = (p: string) =>
-  normalizePath(p)
-    .split('/')
-    .includes('node_modules');
+export const isNodeModulePath = (p: string) => normalizePath(p).split('/').includes('node_modules');
+
+export const getModuleId = (orgImport: string) => {
+  if (orgImport.startsWith('~')) {
+    orgImport = orgImport.substring(1);
+  }
+  const splt = orgImport.split('/');
+  const m = {
+    moduleId: null as string,
+    filePath: null as string,
+    scope: null as string,
+    scopeSubModuleId: null as string,
+  };
+
+  if (orgImport.startsWith('@') && splt.length > 1) {
+    m.moduleId = splt.slice(0, 2).join('/');
+    m.filePath = splt.slice(2).join('/');
+    m.scope = splt[0];
+    m.scopeSubModuleId = splt[1];
+  } else {
+    m.moduleId = splt[0];
+    m.filePath = splt.slice(1).join('/');
+  }
+
+  return m;
+};
 
 export const getPackageDirPath = (p: string, moduleId: string) => {
   const parts = normalizePath(p).split('/');
+  const m = getModuleId(moduleId);
   for (let i = parts.length - 1; i >= 1; i--) {
-    if (parts[i - 1] === 'node_modules' && parts[i] === moduleId) {
-      return parts.slice(0, i + 1).join('/');
+    if (parts[i - 1] === 'node_modules') {
+      if (m.scope) {
+        if (parts[i] === m.scope && parts[i + 1] === m.scopeSubModuleId) {
+          return parts.slice(0, i + 2).join('/');
+        }
+      } else if (parts[i] === m.moduleId) {
+        return parts.slice(0, i + 1).join('/');
+      }
     }
   }
   return null;
